test: raise test coverage for container and task
Add test for checking duplicate names of steps, which raises the test
coverage of container_validation.go to 100%. Replace `ErrInvalidValue`
with `ErrMultipleOneOf` like in the pipeline validation tests for
duplicate tasks.

Add a test for checking of object params without properties and using
an object param in step environment variable value where it is
forbidden.

Signed-off-by: Stanislav Jakuschevskij <stas@two-giants.com>