feature: resolve and create PIP
Extend `ResolvedPipelineTask` data structure to store child
`PipelineRun`s and the resolved pipeline in `ResolvedPipeline` data
structure equal to `ResolvedTask`.

Add `GetPipelineRun` function alias for convenience.

Add condition to `func ResolvePipelineTask` which checks if the given
`PipelineTask` is a `Pipeline` and resolves it by getting the actual
spec. In our "happy path" case it's right on the `PipelineTask` because
we use only `PipelineSpec` and don't need to fetch it from anywhere.

A few downstream methods equal to resolving a `TaskRun` were added.

Extend `resolvePipelineState` with getter for child pipeline runs using
lister and extend `runNextSchedulableTask` with a condition check for a
`PipelineTask` which is a `Pipeline` and implement the creation of
a new `PipelineRun` from the resolved pipeline state and pipeline facts.

Implement child `PipelineRun` status updates on parent `PipelineRun`.

Issues #8760, #7166.

Signed-off-by: Stanislav Jakuschevskij <stas@two-giants.com>