feature: implement e2e PinP using PipelineSpec
[TEP-0056]

Child `PipelineRuns` (PinP) are created by the `PipelineRun` reconciler,
equal to the `TaskRun/CustomRun` implementations.

An event handler for child `PipelineRuns` is registered in controller
entrypoint. This will trigger the reconciliation loop when child
`PipelineRuns` change their state.

Extend `ResolvedPipelineTask` data structure to store child
`PipelineRuns` and the resolved pipeline in `ResolvedPipeline` data
structure equal to `ResolvedTask`.

Add `GetPipelineRun` function alias for convenience.

Add condition to `func ResolvePipelineTask` which checks if the given
`PipelineTask` is a `Pipeline` and resolves it by getting the actual
spec. In our "happy path" case it's right on the `PipelineTask` because
we use `PipelineSpec` and don't need to fetch it from anywhere.

A few downstream methods equal to resolving a `TaskRun` were added.

Extend `resolvePipelineState` with getter for child `PipelineRuns` using
lister and extend `runNextSchedulableTask` with a condition check for a
`PipelineTask` which is a `Pipeline` and implement the creation of
a new `PipelineRun` from the resolved pipeline state and pipeline facts.

To get the status of the child `PipelineRun` success and failure checks
were added to the `PipelineRun` types and extended in the resolution
methods. Test were added to cover the new public methods.

Setting the `ChildReferences` was extended for child `PipelineRuns`.

Issues #8760, #7166.

Signed-off-by: Stanislav Jakuschevskij <stas@two-giants.com>