test: add reconcile child pipeline tests
The test setup is a parent pipeline with one or more embedded
child/grandchild pipelines using the `PipelineSpec` (alpha) field. It
follows the given-when-then test flow arrangement.

The pipeline manifests yaml definitions use variables for every field
which is validated. Multiple helper functions were created equal to
reconciliation unit tests for TaskRuns/CustomRuns:
- `testing.OnePipelineInPipeline`
- `testing.TwoPipelinesInPipelineMixedTasks`
- `testing.NestedPipelinesInPipeline`
- `parse.MustParseChildPipelineRunWithObjectMeta`
- `reconcileOncePinP`
- `validatePinP`
- `getChildPipelineRunsForPipelineRun`
- `getChildPipelineRunByName`
- `validateChildPipelineRunCount`
- `getChildPipelineRunByName`
- `testing.VerifyChildPipelineRunStatusesCount`
- `testing.VerifyChildPipelineRunStatusesNames`
- `testing.childPipelineRunWithObjectMeta`

Test data factory functions were put in the `testing` package in the
`factory.go` file.

The test validates:
- the status and condition of the parent PipelineRuns which should
trigger the creation of the child PipelineRuns,
- the actual created child/grandchild PipelineRuns if they have the correct
metadata i.e. name, owner reference, etc. and the embedded pipelines
from the `PipelineSpec` fields in the pipeline tasks of the parent
pipeline.

Similar checks are performed in `TestReconcile` for `TaskRun` and in
`TestReconcile_V1Beta1CustomTask` for `CustomTasks`.

Issues #8760, #7166.

Signed-off-by: Stanislav Jakuschevskij <stas@two-giants.com>